新包头优化效果对比（3000 人压测） 指标	旧包头 (17B)	新包头 (3-7B)	变化 上行带宽	2.94 MB/s	2.10 MB/s	-29% 下行带宽	10.50 MB/s	9.66 MB/s	-8% 每客户端上行	0.96 KB/s	0.68 KB/s	-29% 每客户端下行	3.42 KB/s	3.14 KB/s	-8% 帧率	20.0 fps	20.0 fps	不变 内存	299 MB	290 MB	-3% 上行省 29%（客户端发的是带 Seq 的小消息，包头从 17B→7B 效果明显）。下行省 8%（推帧包头从 17B→3B，但 payload 158B 占大头）。

+++ cli/Core/FrameSync/FrameSyncProtocol.cs
@@ -8,14 +8,14 @@ namespace BoomNetwork.Core.FrameSync
     /// </summary>
     public static class FrameSyncCmd
     {
-        public const uint SessionBind       = 10;  // 客户端 → 服务器：绑定会话
-        public const uint SessionBindRsp    = 11;  // 服务器 → 客户端：绑定响应
+        public const byte SessionBind       = 10;  // 客户端 → 服务器：绑定会话
+        public const byte SessionBindRsp    = 11;  // 服务器 → 客户端：绑定响应
 
-        public const uint StartFrameSync   = 20;  // 服务器 → 客户端：帧同步开始
-        public const uint StopFrameSync    = 21;  // 服务器 → 客户端：帧同步结束
+        public const byte StartFrameSync   = 20;  // 服务器 → 客户端：帧同步开始
+        public const byte StopFrameSync    = 21;  // 服务器 → 客户端：帧同步结束
 
-        public const uint FrameInput       = 30;  // 客户端 → 服务器：玩家输入
-        public const uint PushFrames       = 31;  // 服务器 → 客户端：推送帧数据
+        public const byte FrameInput       = 30;  // 客户端 → 服务器：玩家输入
+        public const byte PushFrames       = 31;  // 服务器 → 客户端：推送帧数据
     }
 
     /// <summary>
